+++ entity/src/main/java/org/grnet/cat/entities/Comment.java
@@ -1,0 +1,39 @@
+package org.grnet.cat.entities;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "Comment")
+@Getter
+@Setter
+@EqualsAndHashCode(of = "id") // Use the 'id' field for equals and hashCode
+public class Comment {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "assessment_id")
+    @NotNull
+    private Assessment assessment;
+
+    @ManyToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "user_id")
+    @NotNull
+    private User user;
+
+    @Column(name = "created_on")
+    @NotNull
+    private LocalDateTime createdOn;
+
+    @Column(name = "text")
+    @NotNull
+    private String text;
+}